fixed INSANE bug
my dumb ass didn't refactor the whole code after creating the method to check is a variable is imported, so i was check if any import CONTAINED the name of the variable I was checking, meaning that if there was any imported class, every character that makes up the name for that class was considered imported, and therefore, when generating ollir, was using the character like a type, such as :=.i. this wasn't noticed until now because the old tests either didn't have imports, or had io imported but the variables created were usually "a" or any other name.